backend/requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6

# HTTP Clients
httpx==0.26.0
requests==2.31.0
aiohttp==3.9.1  # For async API calls in scripts

# Cloud Storage
boto3==1.34.34  # For Cloudflare R2 (S3-compatible)

# Payment Processing
# Tilopay (no official SDK, using requests)

# Analytics
mixpanel==4.10.1

# Email
sendgrid==6.11.0

# Utils
python-slugify==8.0.1

# Utilities
python-dateutil==2.8.2
python-dotenv==1.0.0

# CORS - Built into FastAPI, no package needed

# Background Tasks (Optional - not required for MVP)
# celery==5.3.6
# redis==5.0.1

# Data Validation
email-validator==2.1.0
